test: mock lib, finalize burn page tests
