+++ jest.config.ts
@@ -133,7 +133,7 @@ const config = {
   // setupFiles: [],
 
   // A list of paths to modules that run some code to configure or set up the testing framework before each test
-  // setupFilesAfterEnv: [],
+  setupFilesAfterEnv: ['<rootDir>/src/test/mocks/setup.ts'],
 
   // The number of seconds after which a test is considered as slow and reported as such in the results.
   // slowTestThreshold: 5,
@@ -178,7 +178,7 @@ const config = {
 
   // A map from regular expressions to paths to transformers
   transform: {
-    '^.+\\.svg$': '<rootDir>/src/__tests__/transformers/svg.js',
+    '^.+\\.svg$': '<rootDir>/src/test/transformers/svg.js',
     "\\.[jt]sx?$": "babel-jest",
   },
