--- src/__tests__/mockFactory.ts
@@ -1,10 +1,0 @@
-import '@testing-library/jest-dom';
-
-jest.mock('@interlay/interbtc-api', () => {
-    const actualInterBtcApi = jest.requireActual('@interlay/interbtc-api');
-
-    return {
-        ...actualInterBtcApi,
-        createInterBtcApi: jest.fn()
-    };
-});
