+++ src/test/mocks/@interlay/interbtc-api/parachain/api.ts
@@ -1,0 +1,3 @@
+const mockApiCreateType = jest.fn(<T extends unknown>(type: string, data: T) => data);
+
+export { mockApiCreateType }
