+++ src/test/mocks/@interlay/interbtc-api/parachain/btcRelay.ts
@@ -1,0 +1,5 @@
+const DEFAULT_BTC_BLOCK_HEIGHT = 1000;
+
+const mockBtcRelayGetLatestBlockHeight = jest.fn(() => DEFAULT_BTC_BLOCK_HEIGHT);
+
+export { mockBtcRelayGetLatestBlockHeight }
