+++ src/test/mocks/@interlay/interbtc-api/parachain/electrsAPI.ts
@@ -1,0 +1,5 @@
+const DEFAULT_BTC_BLOCK_HEIGHT = 1000;
+
+const mockElectrsAPIGetLatestBlockHeight = jest.fn(() => DEFAULT_BTC_BLOCK_HEIGHT);
+
+export { mockElectrsAPIGetLatestBlockHeight }
