+++ src/test/mocks/@interlay/interbtc-api/parachain/index.ts
@@ -1,0 +1,11 @@
+export * from "./api";
+export * from "./btcRelay";
+export * from "./electrsAPI";
+export * from "./fee";
+export * from "./issue";
+export * from "./oracle";
+export * from "./redeem";
+export * from "./system";
+export * from "./tokens";
+export * from "./vaults";
+
