+++ src/test/mocks/@polkadot/constants.ts
@@ -1,0 +1,12 @@
+const DEFAULT_ADDRESS = '5DHdoCL6YPvK5s7N31CdVb3o5bkVGJBCmBZkhXdYzGke5TTd';
+
+const DEFAULT_INJECTED_ACCOUNT = {
+    address: DEFAULT_ADDRESS,
+    meta: {
+        genesisHash: 'test',
+        name: 'test',
+        source: 'test'
+    }
+};
+
+export { DEFAULT_ADDRESS, DEFAULT_INJECTED_ACCOUNT }
