+++ src/test/mocks/intersectionObserver/index.ts
@@ -1,0 +1,9 @@
+const mockIntersectionObserver = jest.fn();
+mockIntersectionObserver.mockReturnValue({
+    observe: jest.fn().mockReturnValue(null),
+    unobserve: jest.fn().mockReturnValue(null),
+    disconnect: jest.fn().mockReturnValue(null)
+});
+window.IntersectionObserver = mockIntersectionObserver;
+
+export { mockIntersectionObserver }
