+++ src/test/mocks/setup.ts
@@ -1,0 +1,17 @@
+import './@interlay/interbtc-api';
+import "./@polkadot/api"
+import "./@polkadot/extension-dapp";
+import "./intersectionObserver";
+import "./fetch"
+
+
+
+
+
+
+
+
+
+
+
+
